Read atlas data
Load all the info needed from projects with (multiple) atlas files.

The engine reads the json files listed in the project using the scon
wrapper.

It creates an Atlas file for each atlas image. These can be subclassed
in the implementations and are returned from the filefactory.

Each ImageFile gets an AtlasFile and atlasframedata set in setAtlasFile
wich can also be overwritten in the subclass if more action is needed.